internal/pkg/scorecard/plugins: make config.go
